Rebalanced rarity and material modifiers
Adjusted rarity and material modifiers to be more consistent. Removed negative stat numbers which could result in undesirable multiplication behavior in certain combinations. Zero is now the worst value for a modifier. Rebalanced gold values to match rarity/material power. Reordered modifiers according to rarity, with the exception of the three worse-than-common rarities.